added wrapper for creating rectangle entity and sprite in basescene

+++ include/Vec2.h
@@ -1,7 +1,7 @@
 #pragma once
 
 struct Vec2 {
-	int x, y;
+	float x, y;
 
 	Vec2 operator+(const Vec2& other) const {
 		return { x + other.x, y + other.y };
@@ -15,7 +15,7 @@ struct Vec2 {
 		return { x * other.x, y * other.y };
 	}
 
-	Vec2 operator*(int s) const {
+	Vec2 operator*(float s) const {
 		return { x * s, y * s };
 	}
